requirements.txt
aiohttp==3.8.1
Django==4.0.3
django_filter==21.1
